Implement text document upload functionality and enhance system robustness
- Added `create_strings_from_file()` and `handle_upload_text()` to support bulk text document processing
- Updated `handle_create_string()` with an embedded mode flag for more flexible string creation
- Increased maximum core size from 1024 to 32768 to support larger graph structures
- Enhanced error handling in `get_node_position()` to return long and provide more detailed error messages
- Updated various functions to handle long node positions and improve type safety
- Added documentation in `cycle-management.md` explaining text document upload process and benefits

+++ src/memory.c
@@ -112,7 +112,7 @@ int unload_node_data(unsigned int node_index) {
     }
     
     // Get node position
-    int position = get_node_position(node_index);
+    long position = get_node_position(node_index);
     if (position < 0) {
         return 0;  // Error already printed by get_node_position
     }
@@ -174,7 +174,7 @@ int load_node_to_core(unsigned int node_index) {
 
 bool check_node_consistency(unsigned int node_index) {
     // Get node position
-    int position = get_node_position(node_index);
+    long position = get_node_position(node_index);
     if (position < 0) return false;
     
     // Open files
